+++ .github/workflows/cd.yml
@@ -1,60 +1,60 @@
-name: CD with Gradle
-
-on:
-  push:
-    branches: [ "master3" ]
-  pull_request:
-    branches: [ "master" ]
-
-
-permissions:
-  contents: read
-
-jobs:
-  build:
-    runs-on: ubuntu-latest
-
-    steps:
-      - uses: actions/checkout@v3
-
-      - name: Set up JDK 17
-        uses: actions/setup-java@v3
-        with:
-          java-version: '17'
-          distribution: "adopt"
-      - name: Make application.properties
-        run: |
-          cd ./src/main/resources
-          touch ./application.properties
-          echo "${{ secrets.PROPERTIES }}" > ./application.properties
-
-        shell: bash
-      - name: Build with Gradle
-        run: |
-          chmod +x ./gradlew
-          ./gradlew clean build -x test
-
-      # dockerfile 을 이미지를 빌드하고, 이를 docker repo 로 push
-      - name: Docker build & push to docker repo
-        run: |
-          docker login -u ${{ secrets.DOCKER_USERNAME }} -p ${{ secrets.DOCKER_PASSWORD }}
-          docker build -f Dockerfile -t harrisbang/schedule:latest .
-          docker push harrisbang/schedule
-
-      # 실행 시, docker-compose를 사용.
-      - name: Deploy to server
-
-        uses: appleboy/ssh-action@master
-        id: deploy
-        with:
-          host: ${{ secrets.SSH_HOST }}
-          username: ec2-user
-          key: ${{ secrets.SSH_KEY }}
-          envs: GITHUB_SHA
-
-          script: |
-            sudo docker rm -f $(docker ps -qa)
-            sudo docker pull harrisbang/schedule:latest 
-            docker image prune -f
-            docker run -d -p 8080:8080 --name schedule harrisbang/schedule
-            
+#name: CD with Gradle
+#
+#on:
+#  push:
+#    branches: [ "master3" ]
+#  pull_request:
+#    branches: [ "master" ]
+#
+#
+#permissions:
+#  contents: read
+#
+#jobs:
+#  build:
+#    runs-on: ubuntu-latest
+#
+#    steps:
+#      - uses: actions/checkout@v3
+#
+#      - name: Set up JDK 17
+#        uses: actions/setup-java@v3
+#        with:
+#          java-version: '17'
+#          distribution: "adopt"
+#      - name: Make application.properties
+#        run: |
+#          cd ./src/main/resources
+#          touch ./application.properties
+#          echo "${{ secrets.PROPERTIES }}" > ./application.properties
+#
+#        shell: bash
+#      - name: Build with Gradle
+#        run: |
+#          chmod +x ./gradlew
+#          ./gradlew clean build -x test
+#
+#      # dockerfile 을 이미지를 빌드하고, 이를 docker repo 로 push
+#      - name: Docker build & push to docker repo
+#        run: |
+#          docker login -u ${{ secrets.DOCKER_USERNAME }} -p ${{ secrets.DOCKER_PASSWORD }}
+#          docker build -f Dockerfile -t harrisbang/schedule:latest .
+#          docker push harrisbang/schedule
+#
+#      # 실행 시, docker-compose를 사용.
+#      - name: Deploy to server
+#
+#        uses: appleboy/ssh-action@master
+#        id: deploy
+#        with:
+#          host: ${{ secrets.SSH_HOST }}
+#          username: ec2-user
+#          key: ${{ secrets.SSH_KEY }}
+#          envs: GITHUB_SHA
+#
+#          script: |
+#            sudo docker rm -f $(docker ps -qa)
+#            sudo docker pull harrisbang/schedule:latest
+#            docker image prune -f
+#            docker run -d -p 8080:8080 --name schedule harrisbang/schedule
+#
