+++ src/test/java/com/sparta/schedules/controllertest/CommentControllerTest.java
@@ -7,8 +7,8 @@ import com.sparta.schedules.domain.comment.dto.CommentRequestDto;
 import com.sparta.schedules.domain.user.entity.User;
 import com.sparta.schedules.domain.user.entity.UserRoleEnum;
 import com.sparta.schedules.global.security.UserDetailsImpl;
-import com.sparta.schedules.domain.comment.service.CommentService;
-import com.sparta.schedules.domain.schedule.service.ScheduleService;
+import com.sparta.schedules.domain.comment.service.CommentServiceImpl;
+import com.sparta.schedules.domain.schedule.service.ScheduleServiceImpl;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -45,9 +45,9 @@ public class CommentControllerTest {
     public MockMvc mvc;
     public Principal mockPrincipal;
     @MockBean
-    CommentService Service;
+    CommentServiceImpl Service;
     @MockBean
-    ScheduleService scheduleService;
+    ScheduleServiceImpl scheduleServiceImpl;
     @Autowired
     public WebApplicationContext context;
     @Autowired
